core/requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
qrcode==7.4.2
Pillow==10.2.0
django-environ==0.11.2 